+++ src/main/java/loc/aliar/monitoringsystem/service/admin/AdminDoctorService.java
@@ -6,5 +6,5 @@ import loc.aliar.monitoringsystem.model.DoctorModel;
 import java.util.List;
 
 public interface AdminDoctorService extends CrudService<Doctor, DoctorModel> {
-    List<DoctorModel> getAllByDepartmentId(Long departmentId);
+    List<DoctorModel> getAllByDepartmentId(Integer departmentId);
 }
